requirements.txt
fastapi
pymongo
playwright
python-dotenv
uvicorn
json